Changed text printed when enemies appear. They are chosen from a list, to get some more flavor and randomness to encounters.
Updated help text function so that help text should be displayed better than before.

Changed inventory, so instead of just a list, now it is a list of dictionaries. Had to refactor for this.

Added functionality: Player can learn skill based on if they have a weapon. This happens(for now) at levelup.

items related:
	player get bonus from items, e.g hit, dmg, armor (str, agi, fort: To be decided..)
	Added function to go through items in inventory and apply the correct stat bonuses to the player.
	If the player has not found a item(weapon / shield) after the scenario is done, you "find" something of the boss.
	added a list of 4 items that you can randomly get at the moment.

Added power attack ability:
	learned if you got a sword
	give you more dmg, but reduces your chance to hit. (not been tested properly yet)

updated scenario dict with more story related information that gets printed to the player.