changed enemy base armor from 8 to 10 (it got a bit too low
changed calculation of enemy xp reward. Xp reward now scales down when enemies are higher level.
Changed xp requirements for levling. They were a bit too low with the previous xp reward calculation. Levling time should scale up so that it takes more battles.

Fixed bug in ending when there were no more scenarios to run.

Changed the print function that makes the text flows over the screen. So that it should try to create new line where it is feasable. So that you don't get loooong lines of text.

Changes in scenario document. To different texts that are printed on the screen.